refactor: 💡 Split ScalaDsl in multiple files
This helps maintanibility but will also allow to change only StepDsl for
Scala 3 whereas other *Dsl can stay as is
