+++ cucumber-scala/src/main/scala/io/cucumber/scala/BaseScalaDsl.scala
@@ -1,0 +1,12 @@
+package io.cucumber.scala
+
+private[scala] trait BaseScalaDsl {
+
+  val NO_REPLACEMENT = Seq[String]()
+  val EMPTY_TAG_EXPRESSION = ""
+  val DEFAULT_BEFORE_ORDER = 1000
+  val DEFAULT_AFTER_ORDER = 1000
+
+  private[scala] val registry: ScalaDslRegistry = new ScalaDslRegistry()
+
+}
